plotter_disk: Drop std::cin.tie and std::ios_base::sync_with_stdio calls
Fixes some weirdness when running tests in CLion. I don't think this
calls are really needed when i read this here https://stackoverflow.com/a/31165481.
If they are for some reason please let me know what they help with because it
seem to be all fine without them and tests in CLion are happy without.