tests: Move `(De)serialization` test case down in hierarchy
This is to work around an issue in the `blake3` source code namely the 
following function having a "init phase" which isn't thread safe 
(Assigning the global variable `g_cpu_features` if its undefined yet):

https://github.com/BLAKE3-team/BLAKE3/blob/4056af6d7ffdf4d13bb776b7ea1db2a6b52d4d75/c/blake3_dispatch.c#L85 

Because its not thread safe it leads to data race which is why the TSAN 
was initially red in this PR since it introduces the `(De)Serialization` 
test which triggeres plot creation with 2 phase1 threads and both 
indirectly trigger `get_cpu_features` concurrently. The workaround here 
is to move the new test below the `F Functions` test which indirectly 
calls `get_cpu_features` from a single thread. This single thread runs 
the "init phase" and since the "init phase" is the non-thread-safe part 
and is only required once the follow up tests are as fine like they were 
before on `main`.